pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "flare-deploy"
version = "0.1.0"
description = "Deploy open-source LLMs on your own cloud infrastructure with automatic cost optimization"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "Flare Contributors" }]
keywords = ["llm", "deployment", "cloud", "mlops", "inference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "httpx>=0.24",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "huggingface_hub>=0.20",
    "aiosqlite>=0.19",
    "apscheduler>=3.10",
    "croniter>=2.0",
]

[project.optional-dependencies]
aws = ["skypilot[aws]>=0.7.0"]
gcp = ["skypilot[gcp]>=0.7.0"]
azure = ["skypilot[azure]>=0.7.0"]
kubernetes = ["skypilot[kubernetes]>=0.7.0"]
all = ["skypilot[aws,gcp,azure,kubernetes]>=0.7.0"]
redis = ["redis>=5.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "httpx>=0.24",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
flare = "flare.cli.main:cli"

[project.urls]
Homepage = "https://github.com/flare-deploy/flare"
Repository = "https://github.com/flare-deploy/flare"
Issues = "https://github.com/flare-deploy/flare/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["flare*"]

[tool.setuptools.package-data]
flare = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
